[Issue-1997] Resolve conflict after merge
