+++ packages/extension-base/src/services/migration-service/scripts/MigrateProvider.ts
@@ -14,7 +14,7 @@ export default abstract class MigrateProvider extends BaseMigrationJob {
     const chainState = state.getChainStateByKey(this.slug);
     const chainInfo = state.getChainInfo(this.slug);
 
-    if (chainState.active && chainState.currentProvider === this.oldProvider) {
+    if (chainState && chainState.active && chainState.currentProvider === this.oldProvider) {
       await state.upsertChainInfo({
         mode: 'update',
         chainEditInfo: {
