+++ packages/extension-koni/public/locales/ja/translation.json
@@ -858,5 +858,7 @@
   "Crowdloan ends on": "",
   "Refunded on": "",
   "Won": "",
-  "In auction": ""
+  "In auction": "",
+  "Token Details": "",
+  "Account Details": ""
 }
